requirements.txt
-e .
pytest
attrs==19.3.0
aws-cdk.aws-cloudwatch==1.23.0
aws-cdk.aws-ec2==1.23.0
aws-cdk.aws-iam==1.23.0
aws-cdk.aws-kms==1.23.0
aws-cdk.aws-ssm==1.23.0
aws-cdk.core==1.23.0
aws-cdk.cx-api==1.23.0
aws-cdk.region-info==1.23.0
cattrs==1.0.0
importlib-resources==1.0.2
jsii==0.21.2
pkg-resources==0.0.0
publication==0.0.3
python-dateutil==2.8.1
six==1.14.0
typing-extensions==3.7.4.1